Substituto cron job for an action on every message
